put usock tools in singularity common
