+++ singularity_common/src/utils/mod.rs
@@ -7,3 +7,4 @@
 pub mod id_map;
 pub mod timer;
 pub mod tree;
+pub mod usock_tools;
