+++ singularity_common/tests/sap_connection_test.rs
@@ -1,9 +1,12 @@
 use events::{ClipboardEvent, CopiedEvent, DragEvent, DraggedEvent, MyEvent, PastedEvent};
 use requests::MyRequest;
-use singularity_common::sap::packet::{
-    universal_client_socket::UniversalClientSocket, universal_server_socket::UniversalServerSocket,
+use singularity_common::{
+    sap::packet::{
+        universal_client_socket::UniversalClientSocket,
+        universal_server_socket::UniversalServerSocket,
+    },
+    utils::usock_tools::{self, UnixServerHost},
 };
-use unix_tools::{ServerHandle, ServerHost};
 
 pub mod events {
     use singularity_common::sap::packet::{IdType, PacketTrait};
@@ -87,75 +90,12 @@ pub mod requests {
     }
 }
 
-/// from usock_demo::connection
-/// TODO: put this in singularity_common
-pub mod unix_tools {
-    use std::os::unix::net::{UnixListener, UnixStream};
-
-    const PATH_PREFIX_ENV_KEY: &str = "XDG_RUNTIME_DIR";
-    const PATH_SUFFIX_ENV_KEY: &str = "SINGULARITY_SERVER";
-
-    /// Server side
-    pub struct ServerHost {
-        path: String,
-        pub listener: UnixListener,
-    }
-    impl ServerHost {
-        pub fn bind_new() -> Option<Self> {
-            let path_prefix = std::env::var(PATH_PREFIX_ENV_KEY).ok()?;
-
-            // TODO
-            let path_suffix = "singularity-0";
-
-            // FIXME, I think this only applies to children processes
-            std::env::set_var(PATH_SUFFIX_ENV_KEY, path_suffix);
-
-            let path = format!("{}/{}", path_prefix, path_suffix);
-            Some(Self {
-                listener: UnixListener::bind(&path).ok()?,
-                path,
-            })
-        }
-    }
-    impl Drop for ServerHost {
-        fn drop(&mut self) {
-            // smh, rust should have some temp_set_env_var function which returns an empty object so it auto removes on drop
-            // std::env::remove_var(PATH_SUFFIX_ENV_KEY);
-            // ^ actually, processes might make this unnecessary
-
-            // unix listener doesn't remove the file on drop
-            if let Err(e) = std::fs::remove_file(&self.path) {
-                dbg!(e);
-            }
-        }
-    }
-
-    /// Client-side
-    pub struct ServerHandle {
-        pub stream: UnixStream,
-    }
-    impl ServerHandle {
-        /// Connect to unix socket at `$XDG_RUNTIME_DIR/$SINGULARITY_SERVER`
-        pub fn connect_from_env() -> Option<Self> {
-            let socket_path = format!(
-                "{}/{}",
-                std::env::var(PATH_PREFIX_ENV_KEY).ok()?,
-                std::env::var(PATH_SUFFIX_ENV_KEY).unwrap_or("singularity-0".to_string())
-            );
-
-            Some(Self {
-                stream: UnixStream::connect(socket_path).ok()?,
-            })
-        }
-    }
-}
-
 /// Currently testing w/ server and client on the same thread and process
 #[test]
 fn sap_connection_test() {
-    let server = ServerHost::bind_new().unwrap();
+    let server = UnixServerHost::bind_new().unwrap();
 
-    let client_side_conn = ServerHandle::connect_from_env().unwrap().stream;
+    let client_side_conn = usock_tools::client_connect_from_env().unwrap();
 
     // blocks until connection (unless you set to non-blocking)
     let (server_side_conn, _address) = server.listener.accept().unwrap();
